fix: Refactor the illustration package'
