+++ draft-packages/illustration/KaizenDraft/Illustration/Players/WebmPlayer.tsx
@@ -1,0 +1,5 @@
+import * as React from "react"
+
+export const Player = () => (
+  <div>TODO</div>
+)
