+++ draft-packages/illustration/KaizenDraft/Illustration/index.ts
@@ -1,3 +1,2 @@
 export * from "./Spot"
 export * from "./Scene"
-export * from "./AnimatedBase"
